refactor: 쿼리스트링을 만들어주는 함수 훅으로 분리
